Optimizations: move common functionality to util.py in preparation for memory operations fusion
